requirements.txt
wheel
argon2-cffi==20.1.0
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-auth==0.1.9
django-cors-headers==3.4.0
django-filter==2.3.0
django-model-utils==4.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-service-objects==0.7.0
django-simple-history==2.11.0
django-storages==1.9.1
django-templated-mail==1.1.1
django-unixtimestampfield==0.4.0
djangorestframework==3.11.0
djangorestframeworkjwt==1.11.0
djoser==2.0.3
easy-thumbnails==2.7
factory-boy==2.12.0
Faker==4.1.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9
